Improvements to the search functionality
Combined all query options together instead of short-circuiting out if a higher priority tier is found.
Added fuzzy-ish search at the end of the query process to search for cards containing the query rather than exactly matching it (if the query is long enough).
Added handling for the 2000-character message limit
Added title + subtitle abbreviations (so Gandalf, Defender of the West will have both dotw and gdotw as valid abbreviations)